Separated database, app init, and server logic
