+++ src/leo-server-app/src/app.ts
@@ -1,0 +1,34 @@
+import express, { Express, Request, Response } from "express";
+import * as dotenv from "dotenv";
+import bodyParser from "body-parser";
+
+dotenv.config({ path: `.env.local`, override: true });
+
+const app: Express = express();
+const cors = require("cors");
+
+const usersRoute = require("./routes/user");
+const { router: satelliteRoute } = require("./routes/satellite");
+const scheduleRoute = require("./routes/schedule");
+const logRoute = require("./routes/log");
+
+// Allow requests from your React app's origin (http://localhost:3000)
+const corsOptions = {
+  origin: "http://localhost:3000",
+};
+
+app.use(cors(corsOptions));
+app.use(bodyParser.urlencoded());
+app.use(bodyParser.json());
+
+// Connect routes to app
+app.use("/users", usersRoute);
+app.use("/satellite", satelliteRoute);
+app.use("/schedule", scheduleRoute);
+app.use("/log", logRoute);
+
+app.get("/", (req: Request, res: Response) => {
+  res.send("Hello, this is Express + TypeScript");
+});
+
+module.exports = app;
