+++ src/leo-server-app/src/tests/satellite.test.ts
@@ -1,5 +1,5 @@
 const request = require("supertest");
-const app = require("../server");
+const app = require("../app");
 const { getSatelliteInfo, setTleLines } = require("../routes/satellite");
 
 let defaultTleLine1 =
